Declare which keys Bouncy Castle accepts.
This declares constraints on which keys Cipher, KeyAgreement, Mac and
Signature instances provided by Bouncy Castle accept. Constraints are
expressed using JCA's SupportedKeyClasses and SupportedKeyFormats
attributes.

The main goal of this change is to avoid having the user of Cipher,
KeyAgreement, Mac and Signature primitives explictly specify which
provider to use. To achieve that each Provider gives JCA sufficient
information to automatically pick the highest priority Provider which
can handle the key with which these primitives are being initialized.
This is better than just picking the highest priority Provider which
implements the requested algorithm, because that Provider might be
unable to handle the key.

When the Provider is specified explicitly, these declarative
constraints are ignored.

This looks messy mostly because of how the JCA's constraining
mechanism works. See
https://docs.oracle.com/javase/8/docs/technotes/guides/security/crypto/HowToImplAProvider.html.
In short, if neither SupportedKeyClasses nor SupportedKeyFormats are
specified on an algorithm, the provider's implementation of the
algorithm is expected to handle arbitary keys. If at least one of
these two constraints is specified, the implementation is expected
to handle keys that match any of the constraints. To match a
constraint, a key need to match any element of the constraint.

Bug: 18088752
Change-Id: I1a2cf862c6eb41510d0ca5237be9ee4bc0220cfe