Register DSA OID for KeyFactory not just Signature
X.509 certificates made with DSA signatures have the X9 DSA with SHA1
OID typically, so we need Bouncycastle to register this OID as an alias
for the DSA KeyFactory.

We also need to remove a manual OID alias added for Signatures which
probably indicates how this slipped through the cracks.

Bug: 21209493
Change-Id: I12a88ead61c626343d96a9c335bdf40e615894bd